moved aggregate to mcts.py
